backend/requirements.txt
flask
joblib
pdfplumber
pillow
pytesseract
scikit-learn
